requirements.txt
fastapi
uvicorn
python-multipart
groq
qdrant-client
sentence-transformers
langchain
langchain-text-splitters
PyMuPDF
numpy
SpeechRecognition
deep-translator
gradio>=5.0
python-dotenv
